Move benchmark into module
