+++ x/bank/bench_test.go
@@ -1,4 +1,4 @@
-package benchmarking
+package bank
 
 import (
 	"testing"
@@ -6,25 +6,12 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/auth"
 	"github.com/cosmos/cosmos-sdk/x/auth/mock"
-	"github.com/cosmos/cosmos-sdk/x/bank"
-	abci "github.com/tendermint/abci/types"
-	crypto "github.com/tendermint/go-crypto"
-)
 
-var (
-	priv1    = crypto.GenPrivKeyEd25519()
-	addr1    = priv1.PubKey().Address()
-	priv2    = crypto.GenPrivKeyEd25519()
-	addr2    = priv2.PubKey().Address()
-	coins    = sdk.Coins{sdk.NewCoin("foocoin", 10)}
-	sendMsg1 = bank.MsgSend{
-		Inputs:  []bank.Input{bank.NewInput(addr1, coins)},
-		Outputs: []bank.Output{bank.NewOutput(addr2, coins)},
-	}
+	abci "github.com/tendermint/abci/types"
 )
 
 func BenchmarkOneBankSendTxPerBlock(b *testing.B) {
-	benchmarkApp, _ := bank.GetBenchmarkMockApp()
+	benchmarkApp, _ := getBenchmarkMockApp()
 
 	// Add an account at genesis
 	acc := &auth.BaseAccount{
